return the real giflib version, rather than the dummy

+++ GIF/imgif.h
@@ -3,7 +3,7 @@
 
 #include "imext.h"
 
-long i_giflib_version(void);
+double i_giflib_version(void);
 i_img *i_readgif_wiol(io_glue *ig, int **colour_table, int *colours);
 i_img *i_readgif_single_wiol(io_glue *ig, int page);
 extern i_img **i_readgif_multi_wiol(io_glue *ig, int *count);
